session: move StreamlinkOptions
- Move `StreamlinkOptions` from `streamlink.session`
  to `streamlink.session.options`
- Move docstring containing options list from `Streamlink.set_option()`
  to `StreamlinkOptions` and add this class to the docs
- Move default options from `Streamlink` constructor
  to `StreamlinkOptions._DEFAULTS`, so all options data is defined
  in the `streamlink.session.options` module
- Move session options tests to `tests.session.test_options`
  and change order of tests
