+++ docs/api/session.rst
@@ -3,3 +3,8 @@ Session
 
 .. autoclass:: streamlink.session.Streamlink
     :member-order: bysource
+
+.. autoclass:: streamlink.session.options.StreamlinkOptions
+
+    .. automethod:: get
+    .. automethod:: set
